refactor: remove babel config since it's no longer needed
