+++ preset.js
@@ -1,7 +1,5 @@
-const babel = require('./dist/cjs/babel');
 const webpackFinal = require('./dist/cjs/webpack');
 
 module.exports = {
-  babel,
   webpackFinal,
 };
